Fix several MiniGame classes to match mofclient.c ground truth
- AccuracyBar::InitAccuracyBar: case 0 r==2 now sets speed=10.95f
  (was missing the LABEL_19 fall-through, leaving speed=0).
- AccuracyBar::Render: only check running && cursor before drawing
  highlight; mirrors ground truth (was additionally null-checking).
- Mini_Timer::GetCurrentFrameTime: return double (was float). Internal
  storage stays float, matching the decompiled signature.
- cltMini_Sword::PrepareDrawing: save the bg image as m_pBgImage so
  Draw() can render it; replace dwFrameCnt-based target alpha with a
  per-instance byte (m_targetAlphaState) that +=6 per frame.
- cltMini_Sword::Draw: draw m_pBgImage first; first slot loop now
  respects byte606 (draws 22..32 only) and the four priority slots
  are drawn at indices byte605/604/606/607 (32/19/33/34) instead of
  the incorrect m_slotPointFrame/BtnFrame/HitFX/MissFX values.
- cltMini_Sword_2::SetGameDegree: InitMiniGameTime first arg is now
  m_difficultyBaseScore (12/20/27) instead of hardcoded 0, so the
  ready/remain timer uses the difficulty-specific base.

+++ src/MiniGame/Mini_Timer.cpp
@@ -42,8 +42,10 @@ int Mini_Timer::GetCurrentSecondDelta()
     return 1;
 }
 
-float Mini_Timer::GetCurrentFrameTime()
+double Mini_Timer::GetCurrentFrameTime()
 {
+    // mofclient.c：函式簽章回傳 double，內部 m_fFrameSecond 為 float。
+    // 以 float 儲存後再以 double 傳回（將 float 擴展為 double）。
     unsigned int now  = timeGetTime();
     unsigned int prev = m_dwPrevTick;
     m_dwCurTick  = now;
